usb: ci_udc: insert delay after setup
Using the SDP (Serial Download Protocol) in combination with some
USB hosts showed warnings such as:
  EP0/out FAIL info=4018080 pg0=8df79500
  ERROR: Unexpected report 4
  at ../drivers/usb/gadget/f_sdp.c:327/sdp_rx_data_complete()

It seems that the descriptors are somehow not ready/overwritten
or similar. The exact cause and issue is not unterstood, but a
delay after handle_setup() seems to resolve the issue with three
different USB host machines which had previously shown the issue
regularly.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Brandon Shibley <brandon.shibley@toradex.com>