colibri_imx7: carve out PMD aligned amount of memory
If the Cortex-M4 is running we carve out 1MiB of memory for
rpmsg communication and/or the firmware running from DDR.
On the Colibri iMX7D this area is in the middle of its 512MiB
of memory. Linux seems not to handle the sparse memory area
well and shows only the first 254MiB of memory as available.
Enabling highmem seems to work around this issue. However,
enabling highmem is not ideal since memory from the highmem
area is more restricted then lowmem.

Aligning the carved out memory to ARM Linux' PMD (page middle
directory) size of 2MiB seem to allow allocating all available
memory as as lowmem. Hence increase the carveout region to
2MiB.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>